# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Book My Cleaning — AI receptionist for Jobber cleaning companies
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quote/{token}:
    get:
      operationId: getPublicQuote
      tags: [publicQuote]
      summary: The customer's own view of their quote
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicQuote"
        "404":
          description: No such quote
  /quote/{token}/approve:
    post:
      operationId: approvePublicQuote
      tags: [publicQuote]
      summary: Customer accepts the quote
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicQuote"
        "404":
          description: No such quote
  /quote/{token}/pay:
    post:
      operationId: payPublicQuote
      tags: [publicQuote]
      summary: Start a Stripe Checkout session for the deposit
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Where to send the customer to pay
          content:
            application/json:
              schema:
                type: object
                required: [checkoutUrl]
                properties:
                  checkoutUrl:
                    type: string
        "400":
          description: Nothing payable on this quote
        "404":
          description: No such quote
        "409":
          description: Already paid
        "502":
          description: Stripe could not be reached
  /quote/{token}/payment/refresh:
    post:
      operationId: refreshPublicQuotePayment
      tags: [publicQuote]
      summary: Re-check payment status with Stripe after checkout
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The quote, with payment status brought up to date
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicQuote"
        "404":
          description: No such quote
  # Public: no authentication. This is where the ad links land — a stranger
  # fills in their details and becomes a row in the Leads inbox.
...
components:
  schemas:
...
    QuoteLineItem:
      type: object
      required: [name, quantity, unitPrice]
      properties:
        name: { type: string, minLength: 1 }
        # Usually hours.
        quantity: { type: number }
        # Negative for a discount line, so no minimum.
        unitPrice: { type: number }
    QuoteTotals:
      type: object
      required:
        [
          lineItems,
          subtotal,
          taxLabel,
          taxRate,
          taxAmount,
          feesLabel,
          feesRate,
          feesAmount,
          total,
          deposit,
        ]
      properties:
        # Rebuilt from the pricing inputs; these are the rows on the estimate.
        lineItems:
          type: array
          items: { $ref: "#/components/schemas/QuoteLineItem" }
        subtotal: { type: number }
        taxLabel: { type: string }
        taxRate: { type: number }
        taxAmount: { type: number }
        feesLabel: { type: string }
        feesRate: { type: number }
        feesAmount: { type: number }
        total: { type: number }
        deposit: { type: number }
        depositEmail: { type: ["string", "null"] }
    # What a customer sees at their quote link. Deliberately narrow: this is
    # served without authentication, so it carries only what belongs on the
    # customer's own estimate and nothing else about the company or its jobs.
    PublicQuote:
      type: object
      required:
        [
          companyName,
          customerName,
          service,
          scheduledForLabel,
          totals,
          approved,
        ]
      properties:
        companyName: { type: string }
        customerName: { type: string }
        customerAddress: { type: ["string", "null"] }
        service: { type: string }
        # The company's own description of this service, if they've written one.
        serviceDescription: { type: ["string", "null"] }
        # Preformatted in the company's timezone — the customer must see the
        # same hour the dispatcher does, not their own device's timezone.
        scheduledForLabel: { type: string }
        totals: { $ref: "#/components/schemas/QuoteTotals" }
        notes: { type: ["string", "null"] }
        sentAtLabel: { type: ["string", "null"] }
        approved: { type: boolean }
        approvedAtLabel: { type: ["string", "null"] }
        # True only once Stripe confirms the money cleared — never on the
        # redirect back from Checkout, which the customer controls.
        depositPaid: { type: boolean }
        depositPaidAmount: { type: ["number", "null"] }
        depositPaidAtLabel: { type: ["string", "null"] }
        # What's still chargeable online, or null when nothing is owed.
        payableAmount: { type: ["number", "null"] }
...